+++ security_id/__init__.py
@@ -1,5 +1,3 @@
-from .validation import (
-    Cusip,
-    Sedol,
-    Isin
-)
+from security_id.ids.sedol import Sedol
+from security_id.ids.cusip import Cusip
+from security_id.ids.isin import Isin
